# docs/swagger.yaml
definitions:
  models.Cliente:
    properties:
      bairro:
        type: string
      cidade:
        type: string
      cpf:
        type: string
      endereco:
        type: string
      id:
        type: integer
      nome:
        type: string
      numero:
        type: string
      tel:
        type: string
      uf:
        type: string
      zap:
        type: string
    type: object
info:
  contact: {}
paths:
  /clientes:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os clientes registrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Cliente'
            type: array
      summary: Lista todos os clientes
      tags:
      - Clientes
    post:
      consumes:
      - application/json
      description: Adiciona um novo cliente no sistema
      parameters:
      - description: Dados do Cliente
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/models.Cliente'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Cliente'
      summary: Cria um novo cliente
      tags:
      - Clientes
  /clientes/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um cliente pelo ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Cliente removido com sucesso
          schema:
            type: string
        "404":
          description: Cliente não encontrado
          schema:
            type: string
      summary: Exclui um cliente
      tags:
      - Clientes
    get:
      consumes:
      - application/json
      description: Retorna os dados de um cliente pelo ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cliente'
        "404":
          description: Cliente não encontrado
          schema:
            type: string
      summary: Busca cliente por ID
      tags:
      - Clientes
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um cliente pelo ID
      parameters:
      - description: ID do Cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados do Cliente
        in: body
        name: cliente
        required: true
        schema:
          $ref: '#/definitions/models.Cliente'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Cliente'
        "404":
          description: Cliente não encontrado
          schema:
            type: string
      summary: Atualiza um cliente
      tags:
      - Clientes
swagger: "2.0"